+++ client/main.js
@@ -9,8 +9,6 @@ STATUS_PENDING=1;
 STATUS_APPROVED=2;
 STATUS_REJECTED=3;
 
-navItems.push('adminMenu');
-
 adminNav = adminNav.concat([
   {
     route: 'posts_pending',
@@ -40,7 +38,6 @@ Deps.autorun(function() {
   }
 });
 
-
 // Sort postModules array position using modulePositions as index
 postModules = _.sortBy(postModules, function(module){return _.indexOf(modulePositions, module.position)});
 
@@ -50,4 +47,11 @@ postMeta = _.sortBy(postMeta, 'order');
 
 Meteor.startup(function () {
   $('#rss-link').attr('title', i18n.t('New Posts'));
+
+  AccountsEntry.config({
+    homeRoute: '/',
+    dashboardRoute: '/',
+    passwordSignupFields: 'USERNAME_AND_EMAIL',
+    profileRoute: 'profile'
+  });
 });