+++ client/views/nav/mobile_nav.js
@@ -1,12 +1,21 @@
 Template[getTemplate('mobile_nav')].helpers({
-  navItems: function () {
-    return navItems;
+  primaryNav: function () {
+    return primaryNav;
+  },
+  secondaryNav: function () {
+    return secondaryNav;
+  },
+  getTemplate: function () {
+    return getTemplate(this);
   },
   canPost: function(){
     return canPost(Meteor.user());
   },
   requirePostsApproval: function(){
     return getSetting('requirePostsApproval');
+  },
+  userMenu: function () {
+    return getTemplate('userMenu');
   }
 });
 
@@ -21,8 +30,13 @@ Template[getTemplate('mobile_nav')].rendered = function () {
 }
 
 Template[getTemplate('mobile_nav')].events({
-  'click .dropdown-top-level': function(e){
+  'click .dropdown-top-level': function (e) {
     e.preventDefault();
     $(e.currentTarget).next().slideToggle('fast');
+  },
+  'click .mobile-nav a': function (e) {
+    if (e.target.className.indexOf('dropdown-top-level') == -1){
+      $('body').removeClass('mobile-nav-open');
+    }
   }
 });