+++ client/views/nav/nav.js
@@ -1,18 +1,30 @@
 Template[getTemplate('nav')].helpers({
-  navItems: function () {
-    return navItems;
+  primaryNav: function () {
+    return primaryNav;
+  },
+  hasPrimaryNav: function () {
+    return !!primaryNav.length;
+  },
+  secondaryNav: function () {
+    return secondaryNav;
+  },
+  hasSecondaryNav: function () {
+    return !!secondaryNav.length;
+  },
+  dropdownClass: function () {
+    return getThemeSetting('useDropdowns', true) ? 'has-dropdown' : 'no-dropdown';
+  },
+  getTemplate: function () {
+    return getTemplate(this);
+  },
+  userMenu: function () {
+    return getTemplate('userMenu');
   },
   site_title: function(){
     return getSetting('title');
   },
   logo_url: function(){
     return getSetting('logoUrl');
-  },
-  logo_height: function(){
-    return getSetting('logoHeight');
-  },
-  logo_width: function(){
-    return getSetting('logoWidth');
   },
   logo_top: function(){
     return Math.floor((70-getSetting('logoHeight'))/2);