+++ collections/users.js
@@ -22,7 +22,17 @@ Schema.User = new SimpleSchema({
   createdAt: {
     type: Date
   },
-  profile: {
+  profile: { // public and modifiable
+    type: Object,
+    optional: true,
+    blackbox: true
+  },
+  data: { // public but not modifiable
+    type: Object,
+    optional: true,
+    blackbox: true
+  },
+  votes: { // used for votes only
     type: Object,
     optional: true,
     blackbox: true