+++ packages/telescope-base/lib/base.js
@@ -31,8 +31,10 @@ adminNav = [];
 // array containing subscriptions to be preloaded
 preloadSubscriptions = [];
 
-// array containing nav items; initialize with views menu
-navItems = ['viewsMenu'];
+// array containing nav items; initialize with views menu and admin menu
+primaryNav = ['viewsMenu', 'adminMenu'];
+
+secondaryNav = ['userMenu', 'notificationsMenu', 'submitButton'];
 
 // object containing post list view parameters
 viewParameters = {}
@@ -161,3 +163,8 @@ getTemplate = function (name) {
   return !!templates[name] ? templates[name] : name;
 }
 
+// ------------------------------ Theme Settings ------------------------------ //
+
+themeSettings = {
+  'useDropdowns': true // whether or not to use dropdown menus in a theme
+};