+++ packages/telescope-lib/lib/lib.js
@@ -3,9 +3,17 @@ getSiteUrl = function () {
 }
 
 getSetting = function(setting, defaultValue){
-  var settings=Settings.find().fetch()[0];
-  if(settings && settings[setting]){
+  var settings = Settings.find().fetch()[0];
+  if(settings && (typeof settings[setting] !== 'undefined')){
     return settings[setting];
+  }else{
+    return typeof defaultValue === 'undefined' ? '' : defaultValue;
+  }
+};
+
+getThemeSetting = function(setting, defaultValue){
+  if(typeof themeSettings[setting] !== 'undefined'){
+    return themeSettings[setting];
   }else{
     return typeof defaultValue === 'undefined' ? '' : defaultValue;
   }
@@ -13,6 +21,10 @@ getSetting = function(setting, defaultValue){
 
 camelToDash = function (str) {
   return str.replace(/\W+/g, '-').replace(/([a-z\d])([A-Z])/g, '$1-$2').toLowerCase();
+}
+
+camelCaseify = function(str) {
+  return dashToCamel(str.replace(' ', '-'));
 }
 
 dashToCamel = function (str) {