+++ packages/telescope-rss/lib/server/rss.js
@@ -1,3 +1,5 @@
+var RSS = Npm.require('rss');
+
 serveRSS = function() {
   var feed = new RSS({
     title: getSetting('title'),
@@ -8,9 +10,10 @@ serveRSS = function() {
   });
   
   Posts.find({status: STATUS_APPROVED}, {sort: {postedAt: -1}, limit: 20}).forEach(function(post) {
+    var description = !!post.body ? post.body+'</br></br>' : '';
     feed.item({
      title: post.title,
-     description: post.body+'</br></br> <a href="'+getPostPageUrl(post._id)+'">Comments</a>',
+     description: description+'<a href="'+getPostUrl(post._id)+'">Discuss</a>',
      author: post.author,
      date: post.postedAt,
      url: getPostLink(post),