+++ packages/telescope-theme-hubble/package.js
@@ -1,17 +1,15 @@
-Package.describe("Telescope Hubble theme");
+Package.describe({summary: "Telescope Hubble theme"});
 
-Package.on_use(function (api) {
+Package.onUse(function (api) {
 
-  // api.use(['telescope-lib'], ['client', 'server']);
-
-  // api.use([
-  //   'jquery',
-  //   'underscore',
-  //   'templating'
-  // ], 'client');
+  api.use(['telescope-lib', 'telescope-base'], ['client', 'server']);
 
   api.add_files([
-    'lib/client/stylesheets/screen.css',
+    'lib/hubble.js',
+    ], ['client', 'server']);
+
+  api.add_files([
+    'lib/client/css/screen.css',
     ], ['client']);
   
 });